Исправления типов

+++ libs/compiler/codegen.c
@@ -841,7 +841,7 @@ void output_export(universal_io *const io, const syntax *const sx)
 {
 	uni_printf(io, "#!/usr/bin/ruc-vm\n");
 
-	uni_printf(io, "%i %i %i %i %i %i %i\n", sx->pc, sx->funcnum, sx->id,
+	uni_printf(io, "%i %i %i %lu %i %i %i\n", sx->pc, sx->funcnum, sx->id,
 				   sx->rp, sx->md, sx->maxdisplg, sx->wasmain);
 
 	for (int i = 0; i < sx->pc; i++)
@@ -862,7 +862,7 @@ void output_export(universal_io *const io, const syntax *const sx)
 	}
 	uni_printf(io, "\n");
 
-	for (int i = 0; i < sx->rp; i++)
+	for (size_t i = 0; i < sx->rp; i++)
 	{
 		uni_printf(io, "%i ", sx->reprtab[i]);
 	}
